fix dockerfile
