+++ lambda/hello/backup.Dockerfile.bak
@@ -1,0 +1,19 @@
+FROM rust:1.69.0
+# 関数名を指定
+ENV FN_NAME hello
+WORKDIR /app
+RUN curl https://lambda.rustup.rs -sSf | sh -s -- -y && \
+    . $HOME/.cargo/env && \
+    cargo install cargo-aws-lambda && \
+    apt-get update && \
+    apt-get install -y zip && \
+    rm -rf /var/lib/apt/lists/*
+COPY . .
+RUN cargo build --release --bin ${FN_NAME}
+RUN cargo test --bin ${FN_NAME} --no-run
+RUN mkdir -p /build && \
+    cd target/release && \
+    cp ${FN_NAME} /build/ && \
+    cd /build && \
+    zip ${FN_NAME}.zip ${FN_NAME}
+CMD ["bash"]
