+++ lambda/hello/src/db/mod.rs
@@ -1,2 +1,2 @@
-pub mod db_repo;
+pub mod impl_db_repo;
 pub mod db_model;
