+++ lambda/hello/src/main.rs
@@ -7,7 +7,7 @@ mod tests;
 pub mod usecase;
 
 use app::repo_trait::DbRepoTrait;
-use db::db_repo::ImplDbRepo;
+use db::impl_db_repo::ImplDbRepo;
 use lambda_http::{run, service_fn, Body, Error, Request, Response};
 use repo_trait::ImplRepoTrait;
 
@@ -17,8 +17,6 @@ use repo_trait::ImplRepoTrait;
 /// - https://github.com/awslabs/aws-lambda-rust-runtime/tree/main/examples
 async fn function_handler(event: Request) -> Result<Response<Body>, Error> {
     // Extract some useful information from the request
-    // dynamodbからデータを挿入する
-    let config = aws_config::from_env().load().await;
     //eventからバイナリデータを取得する
     let bytes = event.body().as_ref();
     // protobufを使う関数をローカルだけでテストするには難しいから一旦デプロイしてデコードができていたらレスポンスにOKと加えて返してクライアントが確認してテスト完了
