+++ lambda/hello/src/repo_trait.rs
@@ -1,4 +1,4 @@
-use crate::{db::db_repo::{ImplDbRepo}, app::repo_trait::DbRepoTrait};
+use crate::{db::impl_db_repo::{ImplDbRepo}, app::repo_trait::DbRepoTrait};
